+++ MyDiet_API/Business/MealRepository.cs
@@ -27,22 +27,22 @@ namespace MyDiet_API.Business
             _logger = logger;
         }
 
-        public async Task<IList<MealDto>> GetAll()
+        public async Task<IList<MealDto>> GetAllAsync()
         {
-            _logger.LogInformation("Entered in GetAll");
+            _logger.LogInformation("Entered in GetAllAsync");
             Stopwatch stopWatch = new Stopwatch();
 
             stopWatch.Start();
-            
+
             List<Meal> meals = await _ctx.Meals.ToListAsync();
-            
+
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
             return _mapper.Map<IList<Meal>, IList<MealDto>>(meals);
         }
 
-        public async Task<MealDto> Get(int id)
+        public async Task<MealDto> GetAsync(int id)
         {
             _logger.LogInformation("Entered in Get with id {}", id);
             Stopwatch stopWatch = new Stopwatch();
@@ -54,18 +54,25 @@ namespace MyDiet_API.Business
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
+
+            if (mealFromDb == null)
+            {
+                _logger.LogError("Meal with id {} not found", id);
+                return null;
+            }
+
             return _mapper.Map<Meal, MealDto>(mealFromDb);
         }
-        
-        public async Task<MealDto> Create(MealDto entity)
+
+        public async Task<MealDto> CreateAsync(MealDto entity)
         {
             _logger.LogInformation("Entered in Create with mealDto {}", entity);
             Stopwatch stopWatch = new Stopwatch();
-            
+
             Meal mealToAdd = _mapper.Map<MealDto, Meal>(entity);
-            
+
             stopWatch.Start();
-            
+
             Meal newMeal = (await _ctx.Meals.AddAsync(mealToAdd)).Entity;
             await _ctx.SaveChangesAsync();
 
@@ -75,14 +82,25 @@ namespace MyDiet_API.Business
             return _mapper.Map<Meal, MealDto>(newMeal);
         }
 
-        public async Task<MealDto> Update(int id, MealDto entity)
+        public async Task<MealDto> UpdateAsync(int id, MealDto entity)
         {
             _logger.LogInformation("Entered in Update with id {}, mealDto {}", id, entity);
             Stopwatch stopWatch = new Stopwatch();
 
             stopWatch.Start();
-            
+
             Meal mealFromDb = await _ctx.Meals.FindAsync(id);
+
+            if (mealFromDb == null)
+            {
+                stopWatch.Stop();
+                _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
+
+                _logger.LogError("Meal with id {} not found", id);
+
+                return null;
+            }
+
             Meal mealToUpdate = _mapper.Map<MealDto, Meal>(entity);
 
             _ctx.Entry(mealFromDb).CurrentValues.SetValues(mealToUpdate);
@@ -94,7 +112,7 @@ namespace MyDiet_API.Business
             return _mapper.Map<Meal, MealDto>(mealToUpdate);
         }
 
-        public async Task Delete(int id)
+        public async Task<bool> DeleteAsync(int id)
         {
             _logger.LogInformation("Entered in Delete with id {}", id);
             Stopwatch stopWatch = new Stopwatch();
@@ -102,15 +120,26 @@ namespace MyDiet_API.Business
             stopWatch.Start();
 
             Meal mealFromDb = await _ctx.Meals.FindAsync(id);
+            if (mealFromDb == null)
+            {
+                stopWatch.Stop();
+                _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
+
+                _logger.LogError("Meal with id {} not found", id);
+
+                return false;
+            }
             _ctx.Meals.Remove(mealFromDb);
 
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
             await _ctx.SaveChangesAsync();
+
+            return true;
         }
 
-        public async Task<bool> AddMealToDiet(int dietId, int mealId)
+        public async Task<bool> AddMealToDietAsync(int dietId, int mealId)
         {
             _logger.LogInformation("Entered in AddMealToDiet with dietId {}, mealId {}", dietId, mealId);
             Stopwatch stopWatch = new Stopwatch();
@@ -118,7 +147,7 @@ namespace MyDiet_API.Business
             stopWatch.Start();
 
             bool exists = await _ctx.DietMeals.AnyAsync(dm => dm.DietId == dietId && dm.MealId == mealId);
-            if(!exists)
+            if (!exists)
             {
                 DietMeal dietMeal = new DietMeal
                 {
@@ -135,11 +164,13 @@ namespace MyDiet_API.Business
                 return true;
             }
 
+            _logger.LogWarning("Meal with id {} is already present in the diet {}", mealId, dietId);
+
             _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
             return false;
         }
 
-        public async Task RemoveMealFromDiet(int dietId, int mealId)
+        public async Task<bool> RemoveMealFromDietAsync(int dietId, int mealId)
         {
             _logger.LogInformation("Entered in RemoveMealFromDiet with dietId {}, mealId {}", dietId, mealId);
             Stopwatch stopWatch = new Stopwatch();
@@ -147,15 +178,28 @@ namespace MyDiet_API.Business
             stopWatch.Start();
 
             DietMeal dietMeal = await _ctx.DietMeals.Where(dm => dm.DietId == dietId && dm.MealId == mealId).FirstOrDefaultAsync();
+            
+            if(dietMeal == null)
+            {
+                stopWatch.Stop();
+                _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
+
+                _logger.LogError("No Meal with id {} is associated with diet {}", mealId, dietId);
+
+                return false;
+            }
+
             _ctx.Remove(dietMeal);
             await _ctx.SaveChangesAsync();
 
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
+
+            return true;
         }
 
-        public async Task<bool> AddProductToMeal(int mealId, int productId)
+        public async Task<bool> AddProductToMealAsync(int mealId, int productId)
         {
             _logger.LogInformation("Entered in AddProductToMeal with mealId {}, productId {}", mealId, productId);
             Stopwatch stopWatch = new Stopwatch();
@@ -163,7 +207,7 @@ namespace MyDiet_API.Business
             stopWatch.Start();
 
             bool exists = await _ctx.MealProducts.AnyAsync(mp => mp.MealId == mealId && mp.ProductId == productId);
-            if(!exists)
+            if (!exists)
             {
                 MealProduct mealProduct = new MealProduct
                 {
@@ -180,13 +224,14 @@ namespace MyDiet_API.Business
                 return true;
             }
 
+            _logger.LogWarning("Product with id {} is already present in the meal {}", productId, mealId);
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
             return false;
         }
 
-        public async Task RemoveProductFromMeal(int mealId, int productId)
+        public async Task<bool> RemoveProductFromMealAsync(int mealId, int productId)
         {
             _logger.LogInformation("Entered in RemoveProductFromMeal with mealId {}, productId {}", mealId, productId);
             Stopwatch stopWatch = new Stopwatch();
@@ -194,6 +239,16 @@ namespace MyDiet_API.Business
             stopWatch.Start();
 
             MealProduct mealProduct = await _ctx.MealProducts.FirstOrDefaultAsync(mp => mp.MealId == mealId && mp.ProductId == productId);
+            
+            if(mealProduct == null)
+            {
+                stopWatch.Stop();
+                _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
+
+                _logger.LogError("No product with id {} is associated with meal {}", productId, mealId);
+
+                return false;
+            }
             _ctx.Remove(mealProduct);
 
             await _ctx.SaveChangesAsync();
@@ -201,6 +256,8 @@ namespace MyDiet_API.Business
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time in {0} ms", stopWatch.ElapsedMilliseconds);
+
+            return true;
         }
 
         public bool CheckIfUnique(string parameter, MealDto entity)
