+++ MyDiet_API/Business/ProductRepository.cs
@@ -27,9 +27,9 @@ namespace MyDiet_API.Business
             _logger = logger;
         }
 
-        public async Task<IList<ProductDto>> GetAll()
+        public async Task<IList<ProductDto>> GetAllAsync()
         {
-            _logger.LogInformation("Entered in GetAll");
+            _logger.LogInformation("Entered in GetAllAsync");
             Stopwatch stopWatch = new Stopwatch();
 
             stopWatch.Start();
@@ -42,7 +42,7 @@ namespace MyDiet_API.Business
             return _mapper.Map<IList<Product>, IList<ProductDto>>(products);
         }
 
-        public async Task<ProductDto> Get(int id)
+        public async Task<ProductDto> GetAsync(int id)
         {
             _logger.LogInformation("Entered in Get with id {}", id);
             Stopwatch stopWatch = new Stopwatch();
@@ -50,17 +50,23 @@ namespace MyDiet_API.Business
             stopWatch.Start();
 
             Product productFromDb = await _ctx.Products.Include(p => p.ProductCategory).FirstOrDefaultAsync(p => p.Id == id);
-            
+
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time is {0} ms", stopWatch.ElapsedMilliseconds);
+
+            if (productFromDb == null)
+            {
+                _logger.LogError("Product with id {} not found", id);
+                return null;
+            }
 
             ProductDto productDto = _mapper.Map<Product, ProductDto>(productFromDb);
 
             return productDto;
         }
 
-        public async Task<ProductDto> Create(ProductDto entity)
+        public async Task<ProductDto> CreateAsync(ProductDto entity)
         {
             _logger.LogInformation("Entered in Create with productDto {}", entity);
             Stopwatch stopWatch = new Stopwatch();
@@ -81,7 +87,7 @@ namespace MyDiet_API.Business
             return _mapper.Map<Product, ProductDto>(newProduct);
         }
 
-        public async Task<ProductDto> Update(int id, ProductDto entity)
+        public async Task<ProductDto> UpdateAsync(int id, ProductDto entity)
         {
             _logger.LogInformation("Entered in Update with id {}, productDto {}", id, entity);
             Stopwatch stopWatch = new Stopwatch();
@@ -89,6 +95,17 @@ namespace MyDiet_API.Business
             stopWatch.Start();
             
             Product productFromDb = await _ctx.Products.Include(p => p.ProductCategory).FirstOrDefaultAsync(p => p.Id == id);
+            
+            if(productFromDb == null)
+            {
+                stopWatch.Stop();
+                _logger.LogInformation("Elapsed time is {0} ms", stopWatch.ElapsedMilliseconds);
+
+                _logger.LogError("Product with id {} not found", id);
+
+                return null;
+            }
+
             Product productToUpdate = _mapper.Map<ProductDto, Product>(entity);
             _ctx.Entry(productFromDb).CurrentValues.SetValues(productToUpdate);
 
@@ -101,13 +118,23 @@ namespace MyDiet_API.Business
             return _mapper.Map<Product, ProductDto>(productToUpdate);
         }
 
-        public async Task Delete(int id)
+        public async Task<bool> DeleteAsync(int id)
         {
             _logger.LogInformation("Entered in Delete with id {}", id);
             Stopwatch stopWatch = new Stopwatch();
 
             stopWatch.Start();
             Product productFromDb = await _ctx.Products.Include(p => p.ProductCategory).FirstOrDefaultAsync(p => p.Id == id);
+            
+            if(productFromDb == null)
+            {
+                stopWatch.Stop();
+                _logger.LogInformation("Elapsed time is {0} ms", stopWatch.ElapsedMilliseconds);
+
+                _logger.LogError("Product with id {} not found", id);
+
+                return false;
+            }
             _ctx.Products.Remove(productFromDb);
 
             await _ctx.SaveChangesAsync();
@@ -115,6 +142,8 @@ namespace MyDiet_API.Business
             stopWatch.Stop();
 
             _logger.LogInformation("Elapsed time is {0} ms", stopWatch.ElapsedMilliseconds);
+
+            return true;
         }
 
         public bool CheckIfUnique(string parameter, ProductDto entity)
